requirements.txt
requests
base64
pynacl